Création de la bibliothèque de class entre l'API et Web
