+++ StageTracker.Shared/ModelsDTO/ApplicationDto.cs
@@ -1,0 +1,3 @@
+﻿namespace StageTracker.Shared.ModelsDTO;
+
+public record ApplicationDto(int Id);
