+++ StageTracker.Shared/ModelsDTO/ClasseDto.cs
@@ -1,0 +1,3 @@
+﻿namespace StageTracker.Shared.ModelsDto;
+
+public record ClasseDto(int Id);
