+++ StageTracker.Shared/ModelsDTO/CompanyDto.cs
@@ -1,0 +1,3 @@
+﻿namespace StageTracker.Shared.ModelsDto;
+
+public record CompanyDto(int Id);
