+++ StageTracker.Shared/ModelsDTO/IntershipDto.cs
@@ -1,0 +1,3 @@
+﻿namespace StageTracker.Shared.ModelsDto;
+
+public record IntershipDto(int Id);
