+++ StageTracker.Shared/ModelsDTO/RemarkDto.cs
@@ -1,0 +1,3 @@
+﻿namespace StageTracker.Shared.ModelsDto;
+
+public record RemarkDto(int Id);
