+++ StageTracker.Shared/ModelsDTO/TeacherDto.cs
@@ -1,0 +1,3 @@
+﻿namespace StageTracker.Shared.ModelsDto;
+
+public record TeacherDto(int Id);
