+++ StageTracker.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StageTrackerAPI", "StageTra
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StageTracker.Web", "StageTracker.Web\StageTracker.Web.csproj", "{0282E805-4F3E-4443-836B-541D39E749F5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "StageTracker.Shared", "StageTracker.Shared\StageTracker.Shared.csproj", "{A1977F6D-81F8-4560-BC99-77523FDA6BCB}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{0282E805-4F3E-4443-836B-541D39E749F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0282E805-4F3E-4443-836B-541D39E749F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0282E805-4F3E-4443-836B-541D39E749F5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A1977F6D-81F8-4560-BC99-77523FDA6BCB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A1977F6D-81F8-4560-BC99-77523FDA6BCB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A1977F6D-81F8-4560-BC99-77523FDA6BCB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A1977F6D-81F8-4560-BC99-77523FDA6BCB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
