--- StageTrackerAPI/ModelsDto/ApplicationDto.cs
@@ -1,3 +1,0 @@
-﻿namespace StageTrackerAPI.ModelsDto;
-
-public record ApplicationDto(int Id);
