--- StageTrackerAPI/ModelsDto/ClasseDto.cs
@@ -1,3 +1,0 @@
-﻿namespace StageTrackerAPI.ModelsDto;
-
-public record ClasseDto(int Id);
