--- StageTrackerAPI/ModelsDto/CompanyDto.cs
@@ -1,3 +1,0 @@
-﻿namespace StageTrackerAPI.ModelsDto;
-
-public record CompanyDto(int Id);
