--- StageTrackerAPI/ModelsDto/IntershipDto.cs
@@ -1,3 +1,0 @@
-﻿namespace StageTrackerAPI.ModelsDto;
-
-public record IntershipDto(int Id);
