--- StageTrackerAPI/ModelsDto/RemarkDto.cs
@@ -1,3 +1,0 @@
-﻿namespace StageTrackerAPI.ModelsDto;
-
-public record RemarkDto(int Id);
