--- StageTrackerAPI/ModelsDto/TeacherDto.cs
@@ -1,3 +1,0 @@
-﻿namespace StageTrackerAPI.ModelsDto;
-
-public record TeacherDto(int Id);
